Move BridgeContext to own file
