+++ lib/exports.js
@@ -26,8 +26,8 @@ module.exports.MatrixRoom = require("./models/rooms/matrix");
 module.exports.RemoteRoom = require("./models/rooms/remote");
 module.exports.StoredEvent = require("./models/events/event");
 
-module.exports.Bridge = require("./bridge").Bridge;
-module.exports.BridgeContext = require("./bridge").BridgeContext;
+module.exports.Bridge = require("./bridge");
+module.exports.BridgeContext = require("./components/bridge-context");
 module.exports.AppServiceRegistration = (
     require("matrix-appservice").AppServiceRegistration
 );
